validation moved to handlers, ErrorList added

+++ backend/src/PetFamily.Application/Volonteers/Delete/HardDeleteVolonteerHandler.cs
@@ -1,5 +1,7 @@
 ﻿using CSharpFunctionalExtensions;
+using FluentValidation;
 using Microsoft.Extensions.Logging;
+using PetFamily.Application.Extensions;
 using PetFamily.Domain.Shared;
 
 namespace PetFamily.Application.Volonteers.Delete
@@ -20,30 +22,39 @@ namespace PetFamily.Application.Volonteers.Delete
             _logger = logger;
         }
 
-        public async Task<Result<Guid, Error>> Handle(
-            DeleteVolonteerRequest request,
+        public async Task<Result<Guid, ErrorList>> Handle(
+            IValidator<DeleteVolonteerCommand> validator,
+            DeleteVolonteerCommand command,
             CancellationToken cancellationToken = default)
         {
-            using var transaction = await _unitOfWork.BeginTransaction(cancellationToken);
+            var validationResult = await validator.ValidateAsync(
+                command,
+                cancellationToken);
+            if (validationResult.IsValid == false)
+                return validationResult.ToErrorList();
+
+            var result = await _volonteersRepository.GetById(command.VolonteerId);
+            if (result.IsFailure)
+                return result.Error.ToErrorList();
+
+            using var transaction = await _unitOfWork.BeginTransactionAsync(cancellationToken);
 
             try
             {
-                var result = await _volonteersRepository.GetById(request.VolonteerId);
-                if (result.IsFailure)
-                    return result.Error;
-
                 await _volonteersRepository.Delete(result.Value);
 
                 transaction.Commit();
 
-                _logger.LogInformation("Volonteer with id {request.VolonteerId) was permenantly deleted}", request.VolonteerId);
+                _logger.LogInformation("Volonteer with id {request.VolonteerId) was permenantly deleted}", command.VolonteerId);
 
                 return result.Value.Id;
             }
             catch (Exception ex)
             {
-                _logger.LogError(ex, "Error occurred while deleting volonteer with id {request.VolonteerId}", request.VolonteerId);
-                return Errors.General.ValueIsInvalid("volonteer deletion");
+                _logger.LogError(ex, "Error occurred while deleting volonteer with id {request.VolonteerId}", command.VolonteerId);
+                return Errors.General
+                    .ValueIsInvalid("volonteer deletion")
+                    .ToErrorList();
             }
         }
     }
